refactor(cli): split pty module into file
